Do not snapshot non-existent outputs for immutable execution
